Day 9

+++ Day09/Day09/Parameter.cs
@@ -5,6 +5,6 @@ namespace Day09
     public class Parameter
     {
         public ParameterMode ParameterMode { get; set; }
-        public int Value { get; set; }
+        public Int64 Value { get; set; }
     }
 }
